backend/pyproject.toml
[tool.poetry]
name = "speaking-english-backend"
version = "0.1.0"
description = "FastAPI backend for the Speaking English App using LM Studio for AI, Whisper for STT and edge-tts for TTS."
authors = ["You <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "0.115.0"}
uvicorn = {extras = ["standard"], version = "0.30.0"}
sqlalchemy = "2.0.36"
alembic = "1.13.2"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
bcrypt = "4.2.1"
httpx = "0.27.2"
python-dotenv = "1.0.1"
asyncpg = "0.29.0"
pydantic-settings = "2.6.1"
faster-whisper = "1.0.3"
requests = "2.32.3"
edge-tts = ">=6.1.10,<8"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.6.0"
poethepoet = "^0.29.0"
psycopg2-binary = "^2.9.9"

[tool.poe.tasks]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
lint = "ruff app"
format = "black app"
create-db = "python scripts/create_pg_db.py"
# Alembic: upgrade head
migrate = "alembic upgrade head"
# Alembic: downgrade one revision
migrate-down = "alembic downgrade -1"
# Alembic: new revision (autogenerate). Custom message: poe migrate-revision -- -m 'your message'
migrate-revision = "alembic revision --autogenerate -m 'update schema'"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

